+++ Source/Core/pointInsideTriangle.js
@@ -1,7 +1,8 @@
 import barycentricCoordinates from "./barycentricCoordinates.js";
 import Cartesian3 from "./Cartesian3.js";
+import defined from "./defined.js";
 
-var coords = new Cartesian3();
+var scratchBarycentricCoords = new Cartesian3();
 
 /**
  * Determines if a point is inside a triangle.
@@ -23,7 +24,16 @@ var coords = new Cartesian3();
  *   new Cesium.Cartesian2(0.0, 1.0));
  */
 function pointInsideTriangle(point, p0, p1, p2) {
-  barycentricCoordinates(point, p0, p1, p2, coords);
+  var result = barycentricCoordinates(
+    point,
+    p0,
+    p1,
+    p2,
+    scratchBarycentricCoords
+  );
+  if (!defined(result)) {
+    return false;
+  }
   return coords.x > 0.0 && coords.y > 0.0 && coords.z > 0;
 }
 export default pointInsideTriangle;